Resolved issues
