+++ tests/test-tokenizer-0-deepseek-coder.cpp
@@ -63,7 +63,7 @@ int main(int argc, char **argv) {
     llama_model * model;
     llama_context * ctx;
 
-    llama_backend_init(false);
+    llama_backend_init();
 
     // load the vocab
     {
